cloud_functions/carefirst-daily-dag-trigger/requirements.txt
google-cloud-storage==1.29.0
google-api-python-client==1.9.0
google-cloud-secret-manager==2.5.0
PGPy==0.5.4
pandas==1.3.0
xlrd==2.0.1
openpyxl==3.0.7
